+++ xscen/extract.py
@@ -5,6 +5,7 @@ import os
 import re
 import warnings
 from collections import defaultdict
+from collections.abc import Sequence
 from copy import deepcopy
 from pathlib import Path
 from typing import Callable, Optional, Union
@@ -655,7 +656,7 @@ def search_data_catalogs(  # noqa: C901
         This does not guarantee that a given warming level will be reached, only that the datasets have corresponding columns in the csv.
         More option can be added by passing a dictionary instead of a boolean.
         If {'ignore_member':True}, it will disregard the member when trying to match the dataset to a column.
-        If {tas_csv: Path_to_csv}, it will use an alternative csv instead of the default one provided by xscen.
+        If {tas_src: Path_to_netcdf}, it will use an alternative netcdf instead of the default one provided by xscen.
         If 'wl' is a provided key, then `xs.get_warming_level` will be called and only datasets that reach the given warming level will be kept.
         This can be combined with other arguments of the function, for example {'wl': 1.5, 'window': 30}.
 
@@ -738,7 +739,6 @@ def search_data_catalogs(  # noqa: C901
         if isinstance(restrict_warming_level, bool):
             restrict_warming_level = {}
         restrict_warming_level.setdefault("ignore_member", False)
-        restrict_warming_level.setdefault("tas_csv", None)
         catalog.esmcat._df = _restrict_wl(catalog.df, restrict_warming_level)
 
     if id_columns is not None or catalog.df["id"].isnull().any():
@@ -927,13 +927,15 @@ def search_data_catalogs(  # noqa: C901
 
 @parse_config
 def get_warming_level(  # noqa: C901
-    realization: Union[xr.Dataset, dict, str, list],
+    realization: Union[
+        xr.Dataset, xr.DataArray, dict, pd.Series, pd.DataFrame, str, list
+    ],
     wl: float,
     *,
     window: int = 20,
-    tas_baseline_period: Optional[list[str]] = None,
+    tas_baseline_period: Optional[Sequence[str]] = None,
     ignore_member: bool = False,
-    tas_csv: Optional[str] = None,
+    tas_src: Optional[Union[str, os.PathLike]] = None,
     return_horizon: bool = True,
 ) -> Union[dict, list[str], str]:
     """
@@ -942,7 +944,7 @@ def get_warming_level(  # noqa: C901
 
     Parameters
     ----------
-    realization : xr.Dataset, dict, str or list of those
+    realization : xr.Dataset, xr.DataArray, dict, str, Series or sequence of those
        Model to be evaluated. Needs the four fields mip_era, source, experiment and member,
        as a dict or in a Dataset's attributes.
        Strings should follow this formatting: {mip_era}_{source}_{experiment}_{member}.
@@ -961,11 +963,12 @@ def get_warming_level(  # noqa: C901
        [start, end] of the base period. The warming is calculated with respect to it. The default is ["1850", "1900"].
     ignore_member : bool
        Decides whether to ignore the member when searching for the model run in tas_csv.
-    tas_csv : str, optional
-       Path to a csv of annual global mean temperature with a row for each year and a column for each dataset.
-       If None, it will default to data/IPCC_annual_global_tas.csv which was built from
+    tas_src : str, optional
+       Path to a netCDF of annual global mean temperature (tas) with an annual "time" dimension
+       and a "simulation" dimension with the following coordinates: "mip_era", "source", "experiment" and "member".
+       If None, it will default to data/IPCC_annual_global_tas.nc which was built from
        the IPCC atlas data from  Iturbide et al., 2020 (https://doi.org/10.5194/essd-12-2959-2020)
-       and extra data from pilot models of MRCC5 and ClimEx.
+       and extra data for missing CMIP6 models and pilot models of CRCM5 and ClimEx.
     return_horizon: bool
         If True, the output will be a list following the format ['start_yr', 'end_yr']
         If False, the output will be a string representing the middle of the period.
@@ -973,10 +976,11 @@ def get_warming_level(  # noqa: C901
     Returns
     -------
     dict, list or str
-        If `realization` is a Dataset, dict or string, the output will follow the format indicated by `return_horizon`.
-        If `realization` is a list, the output will be a dictionary where the keys are the selected columns from the csv
-        and the values follow the format indicated by `return_period`.
+        If `realization` is not a sequence, the output will follow the format indicated by `return_horizon`.
+        If `realization` is a sequence, the output will be a list or dictionary depending on `output`,
+        with values following the format indicated by `return_horizon`.
     """
+    tas_src = tas_src or Path(__file__).parent / "data" / "IPCC_annual_global_tas.nc"
     tas_baseline_period = standardize_periods(
         tas_baseline_period or ["1850", "1900"], multiple=False
     )
@@ -986,13 +990,17 @@ def get_warming_level(  # noqa: C901
 
     FIELDS = ["mip_era", "source", "experiment", "member"]
 
-    if tas_csv is None:
-        tas_csv = Path(__file__).parent / "data" / "IPCC_annual_global_tas.csv"
+    if isinstance(realization, (xr.Dataset, str, dict, pd.Series)):
+        reals = [realization]
+    elif isinstance(realization, pd.DataFrame):
+        reals = (row for i, row in realization.iterrows())
+    elif isinstance(realization, xr.DataArray):
+        reals = realization.values
+    else:
+        reals = realization
 
-    if isinstance(realization, (xr.Dataset, str, dict)):
-        realization = [realization]
     info_models = []
-    for real in realization:
+    for real in reals:
         info = {}
         if isinstance(real, xr.Dataset):
             attrs = get_cat_attrs(real)
@@ -1013,10 +1021,13 @@ def get_warming_level(  # noqa: C901
             ) = real.split("_")
             if ignore_member:
                 info["member"] = ".*"
-        elif isinstance(real, dict) and set(real.keys()).issuperset(
+        # Dict or Series (DataFrame row)
+        elif hasattr(real, "keys") and set(real.keys()).issuperset(
             (set(FIELDS) - {"member"}) if ignore_member else FIELDS
         ):
             info = real
+            if info.get("driving_model") is not None:
+                info["source"] = info["driving_model"]
             if ignore_member:
                 info["member"] = ".*"
         else:
@@ -1025,99 +1036,89 @@ def get_warming_level(  # noqa: C901
             )
         info_models.append(info)
 
-    # open csv, split column names for easier usage
-    annual_tas = pd.read_csv(tas_csv, index_col="year")
-    models = pd.DataFrame.from_records(
-        [c.split("_") for c in annual_tas.columns],
-        index=annual_tas.columns,
-        columns=FIELDS,
-    )
+    # open nc
+    tas = xr.open_dataset(tas_src, engine="h5netcdf").tas
 
-    out = {}
-    for model in info_models:
-        # choose colum based in ds cat attrs
-        mip = models.mip_era.str.fullmatch(model["mip_era"])
-        src = models.source.str.fullmatch(model["source"])
+    def _get_warming_level(model):
+        # choose colum based in ds cat attrs, +'$' to ensure a full match (matches end-of-string)
+        mip = tas.mip_era.str.match(model["mip_era"] + "$")
+        src = tas.source.str.match(model["source"] + "$")
         if not src.any():
-            # Maybe it's an RCM, then source may contain the institute
-            src = models.source.apply(lambda s: model["source"].endswith(s))
-        exp = models.experiment.str.fullmatch(model["experiment"])
-        mem = models.member.str.fullmatch(model["member"])
+            # Maybe it's an RCM, then requested source may contain the institute
+            src = xr.apply_ufunc(model["source"].endswith, tas.source, vectorize=True)
+        exp = tas.experiment.str.match(model["experiment"] + "$")
+        mem = tas.member.str.match(model["member"] + "$")
 
-        candidates = models[mip & src & exp & mem]
-        if candidates.empty:
+        candidates = mip & src & exp & mem
+        if not candidates.any():
             warnings.warn(
-                f"No columns fit the attributes of the input dataset ({model})."
+                f"No simulation fit the attributes of the input dataset ({model})."
             )
-            selected = "_".join([model[c] for c in FIELDS])
-            out[selected] = [None, None] if return_horizon else None
-            continue
-        if len(candidates) > 1:
-            logger.info(
-                "More than one column of the csv fits the dataset metadata. Choosing the first one."
-            )
-        selected = candidates.index[0]
-        right_column = annual_tas.loc[:, selected]
+            return [None, None] if return_horizon else None
 
+        if candidates.sum() > 1:
+            logger.info(
+                "More than one simulation of the database fits the dataset metadata. Choosing the first one."
+            )
+        tas_sel = tas.isel(simulation=candidates.argmax())
+        selected = "_".join([tas_sel[c].item() for c in FIELDS])
         logger.debug(
-            f"Computing warming level +{wl}°C for {model} from column: {selected}."
+            f"Computing warming level +{wl}°C for {model} from simulation: {selected}."
         )
 
-        # compute reference temperature for the warming
-        mean_base = right_column.loc[
-            tas_baseline_period[0] : tas_baseline_period[1]
-        ].mean()
-
-        yearly_diff = right_column - mean_base  # difference from reference
+        # compute reference temperature for the warming and difference from reference
+        yearly_diff = tas_sel - tas_sel.sel(time=slice(*tas_baseline_period)).mean()
 
         # get the start and end date of the window when the warming level is first reached
+        rolling_diff = yearly_diff.rolling(
+            time=window, min_periods=window, center=True
+        ).mean()
         # shift(-1) is needed to reproduce IPCC results.
         # rolling defines the window as [n-10,n+9], but the the IPCC defines it as [n-9,n+10], where n is the center year.
         if window % 2 == 0:  # Even window
-            rolling_diff = (
-                yearly_diff.rolling(window=window, min_periods=window, center=True)
-                .mean()
-                .shift(-1)
-            )
-        else:  # window % 2 == 1:  # Odd windows do not require the shift
-            rolling_diff = yearly_diff.rolling(
-                window=window, min_periods=window, center=True
-            ).mean()
+            rolling_diff = rolling_diff.shift(time=-1)
 
-        yr = rolling_diff.where(rolling_diff >= wl).first_valid_index()
-        if yr is None:
+        yrs = rolling_diff.where(rolling_diff >= wl, drop=True)
+        if yrs.size == 0:
             logger.info(
-                f"Global warming level of +{wl}C is not reached by the last year of the provided 'tas_csv' file for {selected}."
+                f"Global warming level of +{wl}C is not reached by the last year "
+                f"({tas.time[-1].dt.year.item()}) of the provided 'tas_src' database for {selected}."
             )
-            out[selected] = [None, None] if return_horizon else None
-        else:
-            start_yr = int(yr - window / 2 + 1)
-            end_yr = int(yr + window / 2)
-            out[selected] = (
-                standardize_periods([start_yr, end_yr], multiple=False)
-                if return_horizon
-                else str(yr)
-            )
+            return [None, None] if return_horizon else None
 
-    if len(out) != len(realization):
-        warnings.warn(
-            "Two or more input model specifications pointed towards the same column in the CSV, "
-            "the length of the output is different from the input."
+        yr = yrs.isel(time=0).time.dt.year.item()
+        start_yr = int(yr - window / 2 + 1)
+        end_yr = int(yr + window / 2)
+        return (
+            standardize_periods([start_yr, end_yr], multiple=False)
+            if return_horizon
+            else str(yr)
         )
-    if len(realization) == 1:
-        out = out.popitem()[1]
+
+    out = list(map(_get_warming_level, info_models))
+    if isinstance(realization, pd.DataFrame):
+        return pd.Series(out, index=realization.index)
+    if isinstance(realization, xr.DataArray):
+        if return_horizon:
+            return xr.DataArray(
+                out, dims=(realization.dims[0], "bounds"), coords=realization.coords
+            )
+        return xr.DataArray(out, dims=(realization.dims[0],), coords=realization.coords)
+
+    if len(out) == 1:
+        return out[0]
     return out
 
 
 @parse_config
 def subset_warming_level(
     ds: xr.Dataset,
-    wl: float,
+    wl: Union[float, Sequence[float]],
     to_level: str = "warminglevel-{wl}vs{period0}-{period1}",
-    wl_dim: str = "+{wl}Cvs{period0}-{period1}",
+    wl_dim: Union[str, bool] = "+{wl}Cvs{period0}-{period1}",
     **kwargs,
-):
-    """
+) -> Optional[xr.Dataset]:
+    r"""
     Subsets the input dataset with only the window of time over which the requested level of global warming
     is first reached, using the IPCC Atlas method.
 
@@ -1128,65 +1129,159 @@ def subset_warming_level(
        The dataset should include attributes to help recognize it and find its
        warming levels - 'cat:mip_era', 'cat:experiment', 'cat:member', and either
        'cat:source' for global models or 'cat:driving_institution' (optional) + 'cat:driving_model' for regional models.
-    wl : float
+       Or , it should include a `realization` dimension constructed as "{mip_era}_{source or driving_model}_{experiment}_{member}"
+       for vectorized subsetting. Vectorized subsetting is currently only implemented for annual data.
+    wl : float or sequence of floats
        Warming level.
        e.g. 2 for a global warming level of +2 degree Celsius above the mean temperature of the `tas_baseline_period`.
+       Multiple levels can be passed, in which case using "{wl}" in  `to_level` and `wl_dim` is not recommended.
+       Mutliple levels are currently only implemented for annual data.
     to_level :
        The processing level to assign to the output.
        Use "{wl}", "{period0}" and "{period1}" in the string to dynamically include
        `wl`, 'tas_baseline_period[0]' and 'tas_baseline_period[1]'.
-    wl_dim : str
+    wl_dim : str or boolean, optional
        The value to use to fill the new `warminglevel` dimension.
        Use "{wl}", "{period0}" and "{period1}" in the string to dynamically include
        `wl`, 'tas_baseline_period[0]' and 'tas_baseline_period[1]'.
-       If None, no new dimensions will be added.
-
-    kwargs
-        Instructions on how to search for warming levels.
-        The keyword arguments are passed to `get_warming_level()`
-
-        Valid keyword arguments are:
-            window : int
-            tas_baseline_period : list
-            ignore_member : bool
-            tas_csv : str
-            return_horizon: bool
+       If None, no new dimensions will be added, invalid if `wl` is a sequence.
+       If True, the dimension will include `wl` as numbers and units of "degC".
+    \*\*kwargs :
+        Instructions on how to search for warming levels, passed to :py:func:`get_warming_level`.
 
     Returns
     -------
-    xr.Dataset
-        Warming level dataset.
+    xr.Dataset or None
+        Warming level dataset, or None if `ds` can't be subsetted for the requested warming level.
+        The dataset will have a new dimension `warminglevel` with `wl_dim` as coordinates.
+        If `wl` was a list or if ds had a "realization" dim, the "time" axis is replaced
+        by a fake time starting in 1000-01-01 and with a length of `window` years.
+        Start and end years of the subsets are bound in the new coordinate "warminglevel_bounds".
     """
-    start_yr, end_yr = get_warming_level(ds, wl=wl, return_horizon=True, **kwargs)
+    tas_baseline_period = standardize_periods(
+        kwargs.get("tas_baseline_period", ["1850", "1900"]), multiple=False
+    )
+    window = kwargs.get("window", 20)
 
-    if start_yr is None:
-        return None
-    elif any(yr not in ds.time.dt.year for yr in range(int(start_yr), int(end_yr) + 1)):
-        logger.info(
-            f"{ds.attrs.get('cat:id', 'The provided dataset')} does not sufficiently cover the time interval for +{wl}°C ({start_yr}, {end_yr})."
+    # If wl was originally a list, this function is called a 2nd time with a generated fake_time
+    fake_time = kwargs.pop("_fake_time", None)
+    # Fake time generation is needed : real is a dim or multiple levels
+    if (
+        fake_time is None
+        and not isinstance(wl, (int, float))
+        or "realization" in ds.dims
+    ):
+        freq = xr.infer_freq(ds.time)
+        # FIXME: This is because I couldn't think of an elegant way to generate a fake_time otherwise.
+        if not compare_offsets(freq, "==", "YS"):
+            raise NotImplementedError(
+                "Passing multiple warming levels or vectorizing subsetting along the 'realization' dim is currently not supported for non-annual data"
+            )
+        fake_time = xr.date_range(
+            "1000-01-01", periods=window, freq=freq, calendar=ds.time.dt.calendar
         )
-        return None
 
-    # cut the window selected above
-    ds_wl = ds.sel(time=slice(start_yr, end_yr))
+    # If we got a wl sequence, call ourself multiple times and concatenate
+    if not isinstance(wl, (int, float)):
+        if not wl_dim or (isinstance(wl_dim, str) and "{wl}" not in wl_dim):
+            raise ValueError(
+                "`wl_dim` must be True or a template string including '{wl}' if multiple levels are passed."
+            )
+        ds_wl = xr.concat(
+            [
+                subset_warming_level(
+                    ds,
+                    wli,
+                    to_level=to_level,
+                    wl_dim=wl_dim,
+                    _fake_time=fake_time,
+                    **kwargs,
+                )
+                for wli in wl
+            ],
+            "warminglevel",
+        )
+        return ds_wl
 
-    tas_baseline_period = kwargs.get("tas_baseline_period", ["1850", "1900"])
-    if wl_dim:
-        ds_wl = ds_wl.expand_dims(
-            dim={
-                "warminglevel": [
-                    wl_dim.format(
-                        wl=wl,
-                        period0=tas_baseline_period[0],
-                        period1=tas_baseline_period[1],
-                    )
+    # Creating the warminglevel coordinate
+    if isinstance(wl_dim, str):  # a non-empty string
+        wl_crd = xr.DataArray(
+            [
+                wl_dim.format(
+                    wl=wl,
+                    period0=tas_baseline_period[0],
+                    period1=tas_baseline_period[1],
+                )
+            ],
+            dims=("warminglevel",),
+            name="warminglevel",
+        )
+    else:
+        wl_crd = xr.DataArray(
+            [wl], dims=("warminglevel",), name="warminglevel", attrs={"units": "degC"}
+        )
+
+    # For generating the bounds coord
+    date_cls = xc.core.calendar.datetime_classes[ds.time.dt.calendar]
+    if "realization" in ds.dims:
+        # Vectorized subset
+        bounds = get_warming_level(ds.realization, wl, return_horizon=True, **kwargs)
+        reals = []
+        for real in bounds.realization.values:
+            start, end = bounds.sel(realization=real).values
+            if start is not None:
+                data = ds.sel(realization=[real], time=slice(start, end))
+                bnds_crd = [
+                    date_cls(int(start), 1, 1),
+                    date_cls(int(end) + 1, 1, 1) - datetime.timedelta(seconds=1),
                 ]
-            },
-            axis=0,
-        )
-        ds_wl.warminglevel.attrs[
-            "baseline"
-        ] = f"{tas_baseline_period[0]}-{tas_baseline_period[1]}"
+            else:
+                data = (
+                    ds.sel(realization=[real]).isel(time=slice(0, fake_time.size))
+                    * np.NaN
+                )
+                bnds_crd = [np.NaN, np.NaN]
+            reals.append(
+                data.expand_dims(warminglevel=wl_crd).assign_coords(
+                    time=fake_time[: data.time.size],
+                    warminglevel_bounds=(
+                        ("realization", "warminglevel", "bounds"),
+                        [[bnds_crd]],
+                    ),
+                )
+            )
+        ds_wl = xr.concat(reals, "realization")
+    else:
+        # Scalar subset, single level
+        start_yr, end_yr = get_warming_level(ds, wl=wl, return_horizon=True, **kwargs)
+        # cut the window selected above and expand dims with wl_crd
+        ds_wl = ds.sel(time=slice(start_yr, end_yr))
+        if fake_time is None:
+            # WL not reached or completely outside ds time
+            if start_yr is None or ds_wl.time.size == 0:
+                return None
+            ds_wl = ds_wl.expand_dims(warminglevel=wl_crd)
+        else:
+            # WL not reached or not completely inside ds time
+            if start_yr is None or ds_wl.time.size == 0:
+                ds_wl = ds.isel(time=slice(0, fake_time.size)) * np.NaN
+                wlbnds = (("warminglevel", "bounds"), [[np.NaN, np.NaN]])
+            else:
+                wlbnds = (
+                    ("warminglevel", "bounds"),
+                    [
+                        [
+                            date_cls(int(start_yr), 1, 1),
+                            date_cls(int(end_yr) + 1, 1, 1)
+                            - datetime.timedelta(seconds=1),
+                        ]
+                    ],
+                )
+            # We are in an iteration over multiple levels, put the fake time axis, but remember bounds
+            ds_wl = ds_wl.expand_dims(warminglevel=wl_crd).assign_coords(
+                time=fake_time[: ds_wl.time.size],
+                warminglevel_bounds=wlbnds,
+            )
 
     if to_level is not None:
         ds_wl.attrs["cat:processing_level"] = to_level.format(
@@ -1195,6 +1290,13 @@ def subset_warming_level(
             period1=tas_baseline_period[1],
         )
 
+    if not wl_dim:
+        ds_wl = ds_wl.squeeze("warminglevel", drop=True)
+    else:
+        ds_wl.warminglevel.attrs.update(
+            baseline=f"{tas_baseline_period[0]}-{tas_baseline_period[1]}",
+            long_name=f"Warming level for {window}-year periods since {tas_baseline_period[0]}-{tas_baseline_period[1]}",
+        )
     return ds_wl
 
 
@@ -1532,40 +1634,11 @@ def _restrict_wl(df: pd.DataFrame, restrictions: dict):
     df :
         Updated DataFrame.
     """
-    tas_csv = restrictions["tas_csv"]
-    if tas_csv is None:
-        tas_csv = Path(__file__).parent / "data/IPCC_annual_global_tas.csv"
-
-    # open csv
-    annual_tas = pd.read_csv(tas_csv, index_col="year")
-
-    if restrictions["ignore_member"] and "wl" not in restrictions:
-        df["csv_name"] = df["mip_era"].str.cat(
-            [df["source"], df["experiment"]], sep="_"
-        )
-        csv_source = ["_".join(x.split("_")[:-1]) for x in annual_tas.columns[1:]]
-    else:
-        df["csv_name"] = df["mip_era"].str.cat(
-            [df["source"], df["experiment"], df["member"]], sep="_"
-        )
-        csv_source = list(annual_tas.columns[1:])
-
-    if "wl" in restrictions:
-        to_keep = pd.Series(
-            [
-                get_warming_level(x, **restrictions)[0] is not None
-                for x in df["csv_name"]
-            ]
-        )
-    else:
-        to_keep = df["csv_name"].isin(csv_source)
+    restrictions.setdefault("wl", 0)
+    to_keep = get_warming_level(df, return_horizon=False, **restrictions).notnull()
     removed = pd.unique(df[~to_keep]["id"])
-
     df = df[to_keep]
     logger.info(
         f"Removing the following datasets because of the restriction for warming levels: {list(removed)}"
     )
-
-    df = df.drop(columns=["csv_name"])
-
     return df